hyperlynx: split yield adapter to hyperlynx-v3 (protocolId 8107)
Match merged TVL/dimension structure (Hyperlynx V3 = id 8107). V2 yields
deferred until a pair crosses the $10k display threshold.
